// web/src/utils/backend-client/types/providerSetupStepDetails.ts
/**
 * Generated by orval v6.9.6 🍺
 * Do not edit manually.
 * Approvals
 * Granted Approvals API
 * OpenAPI spec version: 1.0
 */
import type { ProviderConfigField } from './providerConfigField';

export interface ProviderSetupStepDetails {
  title: string;
  instructions: string;
  configFields: ProviderConfigField[];
}